Added testing for all version up to 8.2

+++ docker/docker-test-entrypoint.sh
@@ -8,9 +8,12 @@ env | grep VERSION
 echo -n "Installed Wordpress version: "
 wp core version --path=$WORPRESS_ROOT || exit 2
 
-echo "Making a copy of wordpress in the ./mount easier debugging."
-rsync -rclD --delete --exclude storekeeper-for-woocommerce $WORPRESS_DEV_DIR mount/ || exit 3
-
+if [ "$COPY_TO_MOUNT" == "1" ]; then
+  echo "Making a copy of wordpress in the ./mount easier debugging."
+  rsync -rclD --delete --exclude storekeeper-for-woocommerce $WORPRESS_DEV_DIR mount/ || exit 3
+else
+  echo "COPY_TO_MOUNT is not set to 1. No copy"
+fi
 cd $STOREKEEPER_PLUGIN_DIR/tests
 exec $WORPRESS_DEV_DIR/vendor/bin/phpunit .
